Fix definition of per-table file identifiers
The generated headers contain something like this:

  #undef flatbuffers_identifier
  #define flatbuffers_identifier "ABCD"

  #ifndef MyNamespace_MyTable_file_identifier
  #define MyNamespace_MyTable_file_identifier flatbuffers_identifier
  #endif

If several different headers are included, the macro
flatbuffers_identifier is undefined and redefined several times, and at
the end its value it the one from the last included header. Therefore,
all macros <Namespace>_<Table>_file_identifier will have this value,
which is not expected.

Fix this by using the file identifier string instead of the
flatbuffers_identifier macro:

  #ifndef MyNamespace_MyTable_file_identifier
  #define MyNamespace_MyTable_file_identifier "ABCD"
  #endif