Fix per-table file identifiers and extensions (#187)
* Fix definition of per-table file identifiers

The generated headers contain something like this:

  #undef flatbuffers_identifier
  #define flatbuffers_identifier "ABCD"

  #ifndef MyNamespace_MyTable_file_identifier
  #define MyNamespace_MyTable_file_identifier flatbuffers_identifier
  #endif

If several different headers are included, the macro
flatbuffers_identifier is undefined and redefined several times, and at
the end its value it the one from the last included header. Therefore,
all macros <Namespace>_<Table>_file_identifier will have this value,
which is not expected.

Fix this by using the file identifier string instead of the
flatbuffers_identifier macro:

  #ifndef MyNamespace_MyTable_file_identifier
  #define MyNamespace_MyTable_file_identifier "ABCD"
  #endif

* Remove dot from generated file extension

Change the generated file_extension from ".ext" to "ext". This is
more consistent with the schema definition.

* Add definition of per-table file extensions

The documentation says that the file_extension is handled in a similar
manner than file_identifier (see "File Identifiers" in
README.md). However there is currently no per-table file extension.

This commit adds it.