Add Meson build job to CI workflow.
This commit adds a new job to the CI workflow which builds,
runs the tests, and the benchmarks under meson on Linux, Windows
and macos.

The benchmark code up until now was not compiled with the same
flags as the library and runtime and so had avoided some scrutiny
from these warnings. Under Meson this is no longer the case, and
so this code has been fixed up to clean up some warnings