Add FLATCC_BUILD_INTO_SOURCE_DIR cmake option
FLATCC_BUILD_INTO_SOURCE_DIR controls the built destinations of
libraries and executables.

Some flatcc users need both local and cross-compiled versions of the
binaries, so it's helpful to build them into the cmake binary directory
instead of into the source tree.

And when they're built into a directory with Release or Debug path
components, the `_d` suffixes are not necessary.